Moved error functions to separate file. Moved extern crate to lib
