+++ src/lib.rs
@@ -2,6 +2,7 @@ extern crate git2;
 extern crate rustc_serialize;
 extern crate sfml;
 extern crate regex;
+extern crate openssl;
 
 
 pub mod utils;
@@ -10,10 +11,12 @@ mod user;
 mod sequence;
 mod project_types;
 mod error;
+mod err_funcs;
 
 // Re-exports
-pub use error::*;
 pub use init::*;
 pub use user::*;
 pub use sequence::*;
 pub use project_types::*;
+pub use error::*;
+pub use err_funcs::*;
